+++ packages/vite-plugin-simple-hmr/src/transform.ts
@@ -1,5 +1,64 @@
 import type { Node, Program } from "estree";
-import type MagicString from "magic-string";
+import MagicString from "magic-string";
+import { parseAstAsync } from "vite";
+import { name as packageName } from "../package.json";
+
+export async function hmrTransform(code: string) {
+  const magic = new MagicString(code);
+
+  const ast = await parseAstAsync(code);
+  const result = analyzeExports(magic, ast as any);
+
+  if (result.errors.length > 0) {
+    const node = result.errors[0]!.node;
+    const message = "unsupported usage: " + code.slice(node.start, node.end);
+    magic.append("\n" + generateFooterUnsupported(message));
+  } else {
+    magic.append("\n" + generateFooter(result.exportIds));
+  }
+
+  return {
+    code: magic.toString(),
+    map: magic.generateMap(),
+  };
+}
+
+function generateFooter(names: string[]) {
+  const parts = names.map(
+    (name) => `
+  $$registry.exports["${name}"] = {
+    value: ${name},
+    update: ($$next) => {
+      ${name} = $$next;
+    }
+  };
+`
+  );
+
+  // requires dummy "hot.accept" for vite to detect
+  return `
+if (import.meta.env.SSR && import.meta.hot) {
+  const $$hmr = await import("${packageName}/runtime");
+  const $$registry = $$hmr.createRegistry();
+
+${parts.join("\n")}
+
+  $$hmr.setupHot(import.meta.hot, $$registry);
+  import.meta.hot.accept;
+}
+`;
+}
+
+// always invalidate on unsupported usage
+function generateFooterUnsupported(message: string) {
+  return `
+if (import.meta.env.SSR && import.meta.hot) {
+  import.meta.hot.accept(() => {
+    import.meta.hot.invalidate(${JSON.stringify(message)})
+  });
+}
+`;
+}
 
 // traverse export declaration statements based on
 // https://github.com/vitejs/vite/blob/fc2bceb09fb65cc6dc843462f51506586251a703/packages/vite/src/node/ssr/ssrTransform.ts#L172
